requirements.txt
pyinstaller
netifaces
pyyaml
